requirements.txt
requests
pandas
python-whois
ipaddress